MAP-5: Take length limit into account in class name index change set
MySQL has a 255 (usually) char length limit on indexes. Prior to 5.5,
index key length is automatically limited to this length (a 'prefix'
index) but on 5.6 it is an error to try to create an index for a
longer column. Modify the existing change set into a mysql specific
one that limits the index key length and create another change set
for all other dbms.

Note that this commit modifies an existing 1.1.0-SNAPSHOT liquibase
change set. If you have deployed a 1.1.0-SNAPSHOT version from a
prior commit, you may have to drop the index
metadatamapping_idx_mdtm_mdclass manually before deploying this
commit.